refactor(dashboard): change to single naming convention
