+++ src/features/dashboard/index.tsx
@@ -4,7 +4,7 @@ import { Navigate } from "react-router-dom";
 import { selectCurrentUser } from "app/slice/user.slice";
 import { CoursesPath } from "routes/paths";
 
-import DashboardPanel from "./dashboardPanel";
+import DashboardPanel from "./dashboard-panel";
 
 export default function Dashboard() {
   const currentUser = useSelector(selectCurrentUser);
