+++ src/features/dashboard/manage-courses/components/add-course-image.tsx
