+++ src/features/dashboard/manage-courses/index.tsx
@@ -4,7 +4,7 @@ import { useGetInstructorCoursesQuery } from "app/api/courses.api.slice";
 
 import CourseItem from "features/courses/components/course-item";
 import CourseSkeleton from "features/courses/components/course-skeleton";
-import AddCourseImageModal from "./components/modals/addCourseImageModal";
+import AddCourseImageModal from "./components/modals/add-course-image";
 
 import { Course } from "shared/ts/types";
